Moved List to seperate class/file
