+++ src/Radi.js
@@ -14,6 +14,7 @@ import GLOBALS from './consts/GLOBALS';
 import { isWatchable, EMPTY_NODE } from './index';
 import EventService from './EventService';
 import PopulateService from './PopulateService';
+import List from './List';
 import Link from './Link';
 
 export default class Radi {
@@ -163,59 +164,7 @@ export default class Radi {
   }
 
   list(data, act) {
-    if (!data) return '';
-    let link;
-    const fragment = document.createDocumentFragment();
-    const toplink = EMPTY_NODE.cloneNode();
-
-    fragment.appendChild(toplink);
-
-    const cache = data.source[data.prop] || [];
-    const cacheLen = cache.length || 0;
-
-    if (Array.isArray(cache)) {
-      for (let i = 0; i < cacheLen; i++) {
-        fragment.appendChild(act.call(this, cache[i], i));
-      }
-    } else {
-      let i = 0;
-      for (const key in cache) {
-        fragment.appendChild(act.call(this, cache[key], key, i));
-        i++;
-      }
-    }
-
-    link = fragment.lastChild;
-
-    const w = (a, b) => {
-      if (a === 0) return;
-      if (a > 0) {
-        const len = b.length;
-        const start = len - a;
-        for (let i = start; i < len; i++) {
-          fragment.appendChild(act.call(this, b[i], i));
-        }
-        const temp = fragment.lastChild;
-        link.parentElement.insertBefore(fragment, link.nextSibling);
-        link = temp;
-        return;
-      }
-      for (let i = 0; i < Math.abs(a); i++) {
-        const templink = link.previousSibling;
-        link.parentElement.removeChild(link);
-        link = templink;
-      }
-    };
-
-    if (cache.__path) {
-      let len = cacheLen;
-      this.$eventService.on(cache.__path, (e, v) => {
-        w(v.length - len, v);
-        len = v.length;
-      });
-    }
-
-    return fragment;
+    return new List(this, data, act).create();
   };
 
   ll(fn, watch, c) {
