+++ src/__tests__/List.test.js
@@ -1,0 +1,8 @@
+import List from '../List';
+
+// TODO: Test properly
+describe('List.js', () => {
+  it('doesn\'t crash', () => {
+    new List({}, {}, () => {});
+  });
+});
